Organize code and add preview
